- added test for google contacts csv import 0004080: add google contacts csv import definition http://forge.tine20.org/mantisbt/view.php?id=4080
git-svn-id: file:///www/metaways/svn.tine20.org/repo/trunk@19398 8d11cdf3-9d2f-0410-bbb5-e098f187f06b

+++ tests/tine20/Addressbook/Import/CsvTest.php
@@ -4,8 +4,8 @@
  * 
  * @package     Addressbook
  * @license     http://www.gnu.org/licenses/agpl.html
- * @copyright   Copyright (c) 2008-2010 Metaways Infosystems GmbH (http://www.metaways.de)
- * @author      Philipp Schuele <p.schuele@metaways.de>
+ * @copyright   Copyright (c) 2008-2011 Metaways Infosystems GmbH (http://www.metaways.de)
+ * @author      Philipp Schüle <p.schuele@metaways.de>
  * @version     $Id$
  */
 
@@ -13,10 +13,6 @@
  * Test helper
  */
 require_once dirname(dirname(dirname(__FILE__))) . DIRECTORY_SEPARATOR . 'TestHelper.php';
-
-if (!defined('PHPUnit_MAIN_METHOD')) {
-    define('PHPUnit_MAIN_METHOD', 'Addressbook_Import_CsvTest::main');
-}
 
 /**
  * Test class for Addressbook_Import_Csv
@@ -32,6 +28,11 @@ class Addressbook_Import_CsvTest extends PHPUnit_Framework_TestCase
      * @var string $_filename
      */
     protected $_filename = NULL;
+    
+    /**
+     * @var boolean
+     */
+    protected $_deleteImportFile = TRUE; 
     
     /**
      * Runs the test methods of this class.
@@ -64,7 +65,7 @@ class Addressbook_Import_CsvTest extends PHPUnit_Framework_TestCase
     protected function tearDown()
     {
         // cleanup
-        if (file_exists($this->_filename)) {
+        if (file_exists($this->_filename) && $this->_deleteImportFile) {
             unlink($this->_filename);
         }
     }
@@ -105,27 +106,42 @@ class Addressbook_Import_CsvTest extends PHPUnit_Framework_TestCase
     }
     
     /**
+     * import google contacts
+     */
+    public function testImportGoogleContacts()
+    {
+        $this->_filename = dirname(__FILE__) . '/files/google_contacts.csv';
+        $this->_deleteImportFile = FALSE;
+        
+        $result = $this->_doImport(array('dryrun' => TRUE), 'adb_google_import_csv');
+        
+        $this->assertEquals(1, $result['totalcount']);
+        // 6 records fail because they have no org_name / names
+        $this->assertEquals(6, $result['failcount']);
+    }
+    
+    /**
+     * import helper
      * 
      * @param array $_options
+     * @param string $_definitionName
      * @param Addressbook_Model_ContactFilter $_exportFilter
      * @return array
      */
-    protected function _doImport(array $_options, Addressbook_Model_ContactFilter $_exportFilter)
+    protected function _doImport(array $_options, $_definitionName = 'adb_tine_import_csv', Addressbook_Model_ContactFilter $_exportFilter = NULL)
     {
-        $definition = Tinebase_ImportExportDefinition::getInstance()->getByName('adb_tine_import_csv');
+        $definition = Tinebase_ImportExportDefinition::getInstance()->getByName($_definitionName);
         $this->_instance = Addressbook_Import_Csv::createFromDefinition($definition, $_options);
         
         // export first
-        $exporter = new Addressbook_Export_Csv($_exportFilter, Addressbook_Controller_Contact::getInstance());
-        $this->_filename = $exporter->generate();
+        if ($_exportFilter !== NULL) {
+            $exporter = new Addressbook_Export_Csv($_exportFilter, Addressbook_Controller_Contact::getInstance());
+            $this->_filename = $exporter->generate();
+        }
         
         // then import
         $result = $this->_instance->importFile($this->_filename);
         
         return $result;
     }
-}		
-
-if (PHPUnit_MAIN_METHOD == 'Addressbook_Import_CsvTest::main') {
-    Addressbook_Import_CsvTest::main();
 }
